refactor(python-setup): extract shared package-manager signal collector
*Why*
The disk+interpreter signal collection lived privately inside
PackageManagerTelemetry, but the upcoming python-setup visibility gate needs
the exact same signals to decide whether to offer uv-native setup. Duplicating
the collection would let the two drift; extracting it keeps a single-sourced
collection layer.

*What*
- Add language/packageManagerSignals.ts exporting collectPackageManagerSignals
  (projectRoot, env, log?) — the former private collectSignals /
  getInterpreterSource / isUvCreatedVenv / hasActiveCondaInterpreter /
  fileExists / hasRequirementsTxt / readPyproject moved verbatim as guarded
  free functions, with an injectable SignalDebugLog so callers without a logger
  (tests) can omit it and failures still degrade to absent.
- PackageManagerTelemetry now calls collectPackageManagerSignals and drops its
  private copies; behaviour is unchanged.

*Verification*
yarn --cwd packages/databricks-vscode test:unit (305 passing; the unchanged
PackageManagerTelemetry.test.ts suite included).
